argus/APEX-040-25/backend/requirements.txt
Flask==3.0.0
psycopg2-binary==2.9.9
redis==5.0.1
requests==2.31.0
werkzeug==3.0.1
python-dotenv==1.0.0
ipaddress==1.0.23
